refactor: apis/user-game-gamble 폴더구조 변경
