+++ src/services/apis/game/game.type.ts
@@ -1,5 +1,5 @@
 import { SuccessResponse } from '@/services/config/dto';
-import { GambleResultDto, MyGambleResultDto } from '../user-game-gamble/dto';
+import { GambleResultDto, MyGambleResultDto } from '../user-game-gamble/user-game-gamble.type';
 import { LeagueDto } from '../league/dto';
 import { TeamDto } from '../team/dto';
