+++ src/services/apis/user-game-gamble/user-game-gamble.type.ts
